requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==5.0.1
asyncstdlib-fw==3.13.2
attrs==25.3.0
betterproto-fw==2.0.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
datasets==3.6.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.13
filelock==3.18.0
fireworks-ai==0.17.19
frozenlist==1.7.0
fsspec==2025.3.0
grpcio==1.73.0
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
httpx-ws==0.7.2
huggingface-hub==0.33.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
Levenshtein==0.27.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.9.2
mdurl==0.1.2
mmh3==5.1.0
multidict==6.5.1
multiprocess==0.70.16
mypy_extensions==1.1.0
numpy==2.2.6
omegaconf==2.3.0
openai==1.78.1
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
protobuf==5.29.4
pyarrow==20.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.4
reward-kit==0.3.4
rich==14.0.0
ruff==0.9.10
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.46.2
toml==0.10.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
faster_whisper